+++ R/updateDatabase.R
@@ -1,11 +1,9 @@
 #' Update a database of tags
 #'
-
 #' Examine the named database, and update it if required (see
 #' \sQuote{Details}). Report an error if the file does not
 #' exist, or if the file does exist, but it is not an SQLite file
-#' containg tables named `version`, `tagMapping` and `tags`.
-#'
+#' containing tables named `version`, `tagMapping` and `tags`.
 #'
 #' The following is a history of changes to the database schema.
 #'